Simplify dataref structure on fields
The profile type "led" and "data" don't actually need to be different in the way that they declare or use datarefs. For example, we may wish to have a data-only dataref with a condition, such as the `retractable_gear` dataref...

So I've separated the `profile_type` logic from the way we treat the datarefs. Any profile type can now support an `Expr`.