Fix LED states not always changing (#4)
* Fix LED states not always changing

Due to the manipulation of `LED_STATE_CHANGED` we would set it to `true` or `false` depending upon the last LED dataref read.

* Fix race condition in resetting of `LED_STATE_CHANGED`

The race between the goroutine and the `updateLeds` would mean that changes to `LED_STATE_CHANGED` would get lost. This was easy to spot on the Laminar C172 where the LOW VOLTS annunciator would come on _shortly after_ the STARTER ENGAGED annunciator, and would most frequently not light up.